Refactor into rle.gee on top of core rle-python
Split rle-python-gee so the backend-agnostic RLE core lives in the new
rle-python distribution (imports as rle.core) and this package now ships
only the Google Earth Engine backends under the shared rle namespace
(imports as rle.gee).

- Move code from src/rle_python_gee/ to src/rle/gee/ (PEP 420 namespace):
  auth, ee_rle, map, ecosystems (EE backends), aoo (EE backends), upload,
  CLI, and an rle.backends entry point.
- EE backends subclass the core ABCs in rle.core; the auto-detect
  make_ecosystems factory and from_gee_* classmethods are gone in favor of
  explicit construction (GeeEcosystems, AOOGridEEFeatureCollection, ...).
- Core Ecosystems/AOOGrid .to_ee_feature_collection() lazily routes to
  rle.gee.upload.upload_gdf_to_ee_asset.
- Depend on rle-python>=0.2,<0.3; drop deps now provided by core.
- assign_ecosystem_codes CLI moved to rle-python (core functionality).
- Tests: keep EE-only coverage (test_gee_backends.py + ee_* / map / cli);
  local tests migrated to rle-python.
- Update examples and README for the new import paths.
