+++ pyproject.toml
@@ -1,40 +1,42 @@
 [project]
 name = "rle-python-gee"
-version = "0.1.18"
-description = "Tools for IUCN Red List analysis using Google Earth Engine"
+version = "0.2.0"
+description = "Google Earth Engine data-access backends for rle-python (IUCN Red List of Ecosystems)"
 readme = "README.md"
 requires-python = ">=3.11,<3.13"
 dependencies = [
-    "cartopy>=0.24.0",
+    "rle-python>=0.2,<0.3",
     "earthengine-api>=1.6.14",
     "geemap>=0.36.6",
-    "geopandas>=1.1.3",
-    "geoparquet-io",
-    "ipykernel>=7.1.0",
+    "cartopy>=0.24.0",
     "matplotlib>=3.9.0",
-    "numpy>=1.24.0",
-    "pillow>=10.0.0",
-    "pyyaml>=6.0.3",
-    "rasterio>=1.3.0",
     "requests>=2.32.0",
-    "scipy>=1.11.0",
-    "typer>=0.20.0",
     "wkls>=1.0.1",
+    "gcsfs",
+    "pyyaml>=6.0.3",
+    "typer>=0.20.0",
+    "ipykernel>=7.1.0",
 ]
 
 [project.optional-dependencies]
 viz = ["lonboard"]
-gcs = ["gcsfs"]
+
+[project.scripts]
+rle-gee = "rle.gee.cli:app"
+
+[project.entry-points."rle.backends"]
+gee = "rle.gee._entrypoints:register"
 
 [tool.uv.sources]
-geoparquet-io = { git = "https://github.com/geoparquet/geoparquet-io" }
+rle-python = { path = "../rle-python", editable = true }
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/rle_python_gee"]
+only-include = ["src/rle/gee"]
+sources = ["src"]
 
 [dependency-groups]
 dev = [
