+++ tests/test_main.py
@@ -1,19 +1,25 @@
+"""Tests for the rle-python-gee CLI (rle.gee.cli)."""
+
+from unittest.mock import patch
+
 import pytest
 from typer.testing import CliRunner
-from unittest.mock import patch
-from rle_python_gee.main import app
+
+from rle.gee.cli import app
 
 runner = CliRunner()
 
 
+@pytest.mark.unit
 def test_main_no_command():
-    """Test that main app prints the expected message when no command is provided."""
+    """Test that the app prints the expected message when no command is provided."""
     result = runner.invoke(app, [])
     assert result.exit_code == 0
     assert "Hello from rle-python-gee!" in result.stdout
     assert "Use --help to see available commands" in result.stdout
 
 
+@pytest.mark.unit
 def test_main_help():
     """Test that --help flag shows usage information."""
     result = runner.invoke(app, ["--help"])
@@ -22,16 +28,25 @@ def test_main_help():
     assert "test-auth" in result.stdout
 
 
+@pytest.mark.unit
+def test_main_version():
+    """Test that --version prints the version string."""
+    result = runner.invoke(app, ["--version"])
+    assert result.exit_code == 0
+    assert "rle-python-gee version" in result.stdout
 
-@patch('rle_python_gee.main.print_authentication_status')
+
+@pytest.mark.unit
+@patch("rle.gee.cli.print_authentication_status")
 def test_test_auth_command(mock_print_auth):
-    """Test that test-auth command calls print_authentication_status."""
+    """Test that the test-auth command calls print_authentication_status."""
     result = runner.invoke(app, ["test-auth"])
     assert result.exit_code == 0
     assert "Testing Earth Engine authentication..." in result.stdout
     mock_print_auth.assert_called_once()
 
 
+@pytest.mark.unit
 def test_test_auth_help():
     """Test that test-auth command has proper help text."""
     result = runner.invoke(app, ["test-auth", "--help"])
