+++ tests/test_map.py
@@ -12,7 +12,7 @@ from io import BytesIO
 from shapely.geometry import box
 import shapely
 
-from rle_python_gee.map import create_country_map, get_utm_proj_without_limits
+from rle.gee.map import create_country_map, get_utm_proj_without_limits
 
 
 def create_mock_wkb_for_bounds(bounds):
@@ -30,7 +30,7 @@ class TestCreateCountryMap:
         with patch.dict(sys.modules, {'wkls': self.mock_wkls}):
             yield
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_basic_map_creation(self, mock_plt):
         """Test basic map creation with default parameters."""
         # Setup mocks
@@ -54,7 +54,7 @@ class TestCreateCountryMap:
             mock_plt.savefig.assert_called_once()
             mock_plt.close.assert_called_once_with(mock_fig)
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_country_not_found(self, mock_plt):
         """Test that ValueError is raised when country code not found in database."""
         # Mock wkls to raise ValueError when country not found
@@ -108,7 +108,7 @@ class TestCreateCountryMap:
         with pytest.raises(ValueError, match="must be a 2-letter ISO 3166-1 alpha-2 code"):
             create_country_map('U$')
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_valid_lowercase_code(self, mock_plt):
         """Test that lowercase ISO codes are accepted and converted."""
         # Setup mocks
@@ -126,7 +126,7 @@ class TestCreateCountryMap:
             result = create_country_map('sg', output_path)
             assert result == output_path
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_valid_uppercase_code(self, mock_plt):
         """Test that uppercase ISO codes are accepted."""
         # Setup mocks
@@ -144,7 +144,7 @@ class TestCreateCountryMap:
             result = create_country_map('SG', output_path)
             assert result == output_path
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_default_output_path(self, mock_plt):
         """Test that default output path is generated correctly."""
         # Setup mocks
@@ -162,7 +162,7 @@ class TestCreateCountryMap:
         # Should generate 'nz.png'
         assert result == 'nz.png'
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_custom_colors(self, mock_plt):
         """Test map creation with custom fill and edge colors."""
         # Setup mocks
@@ -195,7 +195,7 @@ class TestCreateCountryMap:
             assert call_kwargs['edgecolor'] == 'darkred'
             assert call_kwargs['linewidth'] == 2.5
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_no_border(self, mock_plt):
         """Test map creation with show_border=False."""
         # Setup mocks
@@ -224,7 +224,7 @@ class TestCreateCountryMap:
             # Verify add_geometries was called
             mock_ax.add_geometries.assert_not_called()
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_spines_hidden(self, mock_plt):
         """Test that plot frame spines are hidden."""
         # Setup mocks
@@ -246,7 +246,7 @@ class TestCreateCountryMap:
             for spine in mock_ax.spines.values():
                 spine.set_visible.assert_called_once_with(False)
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_custom_title(self, mock_plt):
         """Test map creation with custom title."""
         # Setup mocks
@@ -267,7 +267,7 @@ class TestCreateCountryMap:
             # Verify title was set
             mock_plt.title.assert_called_once_with('Kenya Wildlife', fontsize=16, fontweight='bold')
 
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.plt')
     def test_no_title(self, mock_plt):
         """Test map creation with empty title."""
         # Setup mocks
@@ -298,8 +298,8 @@ class TestEarthEngineBasemap:
         with patch.dict(sys.modules, {'wkls': self.mock_wkls}):
             yield
 
-    @patch('rle_python_gee.map.requests.get')
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.requests.get')
+    @patch('rle.gee.map.plt')
     def test_ee_image_basemap(self, mock_plt, mock_requests):
         """Test map creation with Earth Engine image basemap."""
         # Setup mocks
@@ -333,7 +333,7 @@ class TestEarthEngineBasemap:
             mock_mask.getDownloadURL.return_value = 'http://example.com/mask.png'
             mock_ee_image.mask.return_value = mock_mask
 
-            with patch('rle_python_gee.map.ee') as mock_ee:
+            with patch('rle.gee.map.ee') as mock_ee:
                 # Mock ee.Projection and ee.Geometry.Rectangle
                 mock_ee.Projection.return_value = Mock()
                 mock_ee.Geometry.Rectangle.return_value = Mock()
@@ -353,8 +353,8 @@ class TestEarthEngineBasemap:
                 # Verify imshow was called to display the basemap
                 mock_ax.imshow.assert_called_once()
 
-    @patch('rle_python_gee.map.requests.get')
-    @patch('rle_python_gee.map.plt')
+    @patch('rle.gee.map.requests.get')
+    @patch('rle.gee.map.plt')
     def test_ee_image_clipped(self, mock_plt, mock_requests):
         """Test map creation with clipped Earth Engine image."""
         # Setup mocks
@@ -387,7 +387,7 @@ class TestEarthEngineBasemap:
             mock_clipped.getDownloadURL.return_value = 'http://example.com/image.png'
             mock_ee_image.clip.return_value = mock_clipped
 
-            with patch('rle_python_gee.map.ee') as mock_ee:
+            with patch('rle.gee.map.ee') as mock_ee:
                 # Mock ee.Projection, ee.Geometry, and ee.Geometry.Rectangle
                 mock_ee.Projection.return_value = Mock()
                 mock_ee_geom = Mock()
